addressed code comments
